chore: restructure wip
